requirements.txt
amqp==2.2.0
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
billiard==3.5.0.2
celery==4.0.2
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
constantly==15.1.0
cryptography==1.9
cssselect==1.0.1
DateTime==4.2
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.3
django-celery-beat==1.0.1
django-toolbelt==0.0.1
djangorestframework==3.6.3
enum34==1.1.6
gunicorn==19.7.1
hyperlink==17.2.1
idna==2.5
incremental==17.5.0
inflection==0.3.1
ipaddress==1.0.18
kombu==4.0.2
lxml==3.8.0
more-itertools==3.2.0
numpy==1.13.1
pandas==0.20.3
parsel==1.2.0
psycopg2==2.7.1
pyasn1==0.2.3
pyasn1-modules==0.0.9
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.1.0
python-dateutil==2.6.1
pytz==2017.2
Quandl==3.2.0
queuelib==1.4.2
requests==2.17.3
scikit-learn==0.18.2
scipy==0.19.1
Scrapy==1.4.0
scrapy-djangoitem==1.1.1
service-identity==17.0.0
six==1.10.0
static3==0.7.0
Twisted==17.5.0
urllib3==1.21.1
vine==1.1.3
w3lib==1.17.0
zope.interface==4.4.2
